test(webapp): split runsReplicationService part5 + real measured shard weights
