+++ test-timings.json
@@ -53,13 +53,13 @@
   "apps/webapp/test/replay-after-crash.test.ts": 2233,
   "apps/webapp/test/runsBackfiller.test.ts": 15478,
   "apps/webapp/test/runsReplicationBenchmark.test.ts": 0,
-  "apps/webapp/test/runsRepository.part1.test.ts": 46000,
-  "apps/webapp/test/runsRepository.part2.test.ts": 46000,
+  "apps/webapp/test/runsRepository.part1.test.ts": 53000,
+  "apps/webapp/test/runsRepository.part2.test.ts": 57000,
   "apps/webapp/test/sanitizeRowsOnParseError.test.ts": 8,
   "apps/webapp/test/sentryTenantContext.test.ts": 5,
   "apps/webapp/test/sentryTraceContext.server.test.ts": 12,
   "apps/webapp/test/sessionDuration.test.ts": 18416,
-  "apps/webapp/test/sessionsReplicationService.test.ts": 32887,
+  "apps/webapp/test/sessionsReplicationService.test.ts": 30000,
   "apps/webapp/test/shouldRevalidateRunsList.test.ts": 5,
   "apps/webapp/test/slackErrorAlerts.test.ts": 0,
   "apps/webapp/test/tenantContext.test.ts": 26,
@@ -74,7 +74,7 @@
   "apps/webapp/test/workerQueueSplit.test.ts": 3,
   "apps/webapp/test/components/DateTime.test.ts": 24,
   "apps/webapp/test/engine/batchPayloads.test.ts": 5018,
-  "apps/webapp/test/engine/streamBatchItems.test.ts": 47000,
+  "apps/webapp/test/engine/streamBatchItems.test.ts": 45000,
   "apps/webapp/test/engine/taskIdentifierRegistry.test.ts": 13152,
   "apps/webapp/test/engine/triggerTask.test.ts": 31630,
   "apps/webapp/test/presenters/mapRunToLiveFields.test.ts": 3,
@@ -209,12 +209,13 @@
   "internal-packages/clickhouse/src/client/client.test.ts": 9138,
   "internal-packages/sdk-compat-tests/src/tests/bundler.test.ts": 348,
   "internal-packages/sdk-compat-tests/src/tests/import.test.ts": 4742,
-  "apps/webapp/test/runsReplicationService.part1.test.ts": 73000,
+  "apps/webapp/test/runsReplicationService.part1.test.ts": 74000,
   "apps/webapp/test/runsReplicationService.part2.test.ts": 64000,
-  "apps/webapp/test/runsReplicationService.part3.test.ts": 43000,
-  "apps/webapp/test/runsReplicationService.part4.test.ts": 61000,
+  "apps/webapp/test/runsReplicationService.part3.test.ts": 30000,
+  "apps/webapp/test/runsReplicationService.part4.test.ts": 70000,
   "apps/webapp/test/runsReplicationService.part5.test.ts": 43000,
-  "apps/webapp/test/runsReplicationService.part6.test.ts": 43000,
-  "apps/webapp/test/runsRepository.part3.test.ts": 34000,
-  "apps/webapp/test/runsRepository.part4.test.ts": 46000
+  "apps/webapp/test/runsReplicationService.part6.test.ts": 32000,
+  "apps/webapp/test/runsRepository.part3.test.ts": 43000,
+  "apps/webapp/test/runsRepository.part4.test.ts": 57000,
+  "apps/webapp/test/runsReplicationService.part7.test.ts": 43000
 }
